+++ CHANGELOG.md
@@ -35,6 +35,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Gitlab CI
 - Localization `mainet` -> `mainnet`
 
+#### Removed
+- Module `@intlify/vue-i18n-loader`
+
 ## [0.1.0-rc.3] - 2021-10-19
 #### Added
 - Auctions tab shell